Move import commands under import category
